Consider 'removed' files as changes worth a commit.
This fixes the behavior of `revert when applied to a change that only
introduced a new file.

What used to happen before was:

- A change introduces a new file "foo".

- `revert is called on the change.

- lib/revert executes "hg revert", which removes (instead of just
  deleting) the file "foo".

- transact() uses "hg status -umad", which does not consider the file
  removal a change, and therefore makes no commit. This leaves the
  repository in an inconsistent state, but to a casual observer the
  `revert seems to have worked, as the file has been deleted from the
  working copy.

- A new, unrelated change is made.

- transact()'s "hg status -umad" check notices the new change,
  acquires the lock and cleans up the working directory with "hg up
  -C". This has the effect of discarding the uncommitted `revert,
  inadvertently restoring the removed file.

As a result, while `revert of a file addition initially appears to
work, no commit is made and the next commit actually adds the file
back again.

This commit just changes the "hg status" argument from -umad to -rumad
(which happens to be the default). As far as I can tell, this should
not cause any undesired side effects, as commands in the sandbox can
never actually remove files (they can just delete them). Though I have
to admit I'm not entirely sure why "-umad" was used in the first
place.

(Code archeology places the introduction of this to the initial 2011
commits intoducing transactions in lib/server.)

--HG--
branch : revert-fixup