Avoid running `fetch and `revert calls twice.
The transact() logic of running with a shared lock, then cleaning up
and running again with the exclusive lock is not really that
well-suited to the `fetch and `revert built-in.

Especially for `fetch, it results in every remote file being
downloaded twice. For `revert it's not really a problem, but not that
useful either.

This commit just skips over the shared call for `fetch and `revert.

--HG--
branch : revert-fixup