Avoid the find-rm '*.orig' step by disabling 'hg revert' backups.
Existing `revert code has a manual cleanup step that finds and removes
the "foo.orig" backup files 'hg revert' uses. While this mostly works,
it causes spurious error messages when a *directory* matching that
pattern exists (or has ever existed, as the find in the
repository. This is currently the case for the #esoteric HackEgo,
which means every `revert looks like:

17-10-2016 01:39:00 <oerjan> `revert
17-10-2016 01:39:13 <HackEgo> rm: cannot remove `/home/hackbot/hackbot.hg/multibot_cmds/env/.hg/store/data/canary.orig': Is a directory \ Done.

'hg revert' has a --no-backup option that disables those files, which
makes the cleanup step unnecessary. Let's use it for `revert.

(Given that the find descends to .hg, it might even cause some
non-cosmetic problems if the file it finds wasn't a directory.)

--HG--
branch : revert-fixup