Restore working copy if commit blocked by canary.
While the removal of the canary file does block a commit from being
made, the working copy is not actually restored, leaving the
repository in an inconsistent state. For example, after a "rm -rf *",
the working copy will be entirely empty.

This is not normally noticeable[*], as the next transaction will see
the broken files, think they were caused by the current command, and
restore the working copy. But it's still inelegant.

Avoid this by doing the same "remove broken files, call 'hg up"
cleanup we do between the shared-lock and exclusive-lock iterations
whenever a commit is skipped due to a missing canary.

[*] Barring tricks via non-persistent state through .hgignore.

--HG--
branch : revert-fixup